Support for collection aliases.
The collection aliases are present under:

/org/freedesktop/secrets/aliases/xxxx

So the default and network collections would be usable via
the object path, in addition to their normal object paths:

/org/freedesktop/secrets/aliases/default
/org/freedesktop/secrets/aliases/network

To manage the aliases, two new methods are added to the
Service interface: ReadAlias() and SetAlias(). BTW, These
are not properties on Collection because a collection can
be have multiple aliases pointing to it.