Add _NET_WM_STATE_FOCUSED
Patch from Rui Tiago Cação Matos to add a a new _NET_WM_STATE atom that
accurately tracks when the window manager shows the window decorations
as active or inactive. This allows for better handling of cases where
modal dialogs might be shown as a unit with a parent window, or where
grabs cause a disjunction between X input focus and the focus displayed
to the user.