desktop-entry: Restore support for Desktop Actions
Reintroduce the Actions key, which was removed around version 1.1,
and give it a formal specification, detailing what actions applications
are expected to provide and who should make use of this information.
Names were taken from the original specification, since apparently
they have been in use during all this time.

Also remove X-KDE-* keys from the example, as it's unrelated to the
spec.